+++ FlutterChatAppTutorial/lib/core/shared_preferences/helperfunctions.dart
@@ -1,43 +1,42 @@
 import 'package:shared_preferences/shared_preferences.dart';
 
-class HelperFunctions{
-
+class HelperFunctions {
   static String sharedPreferenceUserLoggedInKey = "ISLOGGEDIN";
   static String sharedPreferenceUserNameKey = "USERNAMEKEY";
   static String sharedPreferenceUserEmailKey = "USEREMAILKEY";
 
   /// saving data to sharedpreference
-  static Future<bool> saveUserLoggedInSharedPreference(bool isUserLoggedIn) async{
-
+  static Future<bool> saveUserLoggedInSharedPreference(
+      bool isUserLoggedIn) async {
     SharedPreferences preferences = await SharedPreferences.getInstance();
-    return await preferences.setBool(sharedPreferenceUserLoggedInKey, isUserLoggedIn);
+    return await preferences.setBool(
+        sharedPreferenceUserLoggedInKey, isUserLoggedIn);
   }
 
-  static Future<bool> saveUserNameSharedPreference(String userName) async{
+  static Future<bool> saveUserNameSharedPreference(String userName) async {
     SharedPreferences preferences = await SharedPreferences.getInstance();
     return await preferences.setString(sharedPreferenceUserNameKey, userName);
   }
 
-  static Future<bool> saveUserEmailSharedPreference(String userEmail) async{
+  static Future<bool> saveUserEmailSharedPreference(String userEmail) async {
     SharedPreferences preferences = await SharedPreferences.getInstance();
     return await preferences.setString(sharedPreferenceUserEmailKey, userEmail);
   }
 
   /// fetching data from sharedpreference
 
-  static Future<bool> getUserLoggedInSharedPreference() async{
+  static Future<bool> getUserLoggedInSharedPreference() async {
     SharedPreferences preferences = await SharedPreferences.getInstance();
     return await preferences.getBool(sharedPreferenceUserLoggedInKey);
   }
 
-  static Future<String> getUserNameSharedPreference() async{
+  static Future<String> getUserNameSharedPreference() async {
     SharedPreferences preferences = await SharedPreferences.getInstance();
     return await preferences.getString(sharedPreferenceUserNameKey);
   }
 
-  static Future<String> getUserEmailSharedPreference() async{
+  static Future<String> getUserEmailSharedPreference() async {
     SharedPreferences preferences = await SharedPreferences.getInstance();
     return await preferences.getString(sharedPreferenceUserEmailKey);
   }
-
 }
