+++ FlutterChatAppTutorial/lib/feature/data/models/user.dart
