+++ FlutterChatAppTutorial/lib/feature/data/services/auth.dart
@@ -1,9 +1,7 @@
-import 'package:chatapp/helper/helperfunctions.dart';
-import 'package:chatapp/models/user.dart';
-import 'package:chatapp/views/chat.dart';
+import 'package:chatapp/core/shared_preferences/helperfunctions.dart';
+import 'package:chatapp/feature/data/models/user.dart';
+
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/material.dart';
-import 'package:google_sign_in/google_sign_in.dart';
 
 class AuthService {
   final FirebaseAuth _auth = FirebaseAuth.instance;
@@ -42,27 +40,6 @@ class AuthService {
     } catch (e) {
       print(e.toString());
       return null;
-    }
-  }
-
-  Future<FirebaseUser> signInWithGoogle(BuildContext context) async {
-    final GoogleSignIn _googleSignIn = new GoogleSignIn();
-
-    final GoogleSignInAccount googleSignInAccount =
-        await _googleSignIn.signIn();
-    final GoogleSignInAuthentication googleSignInAuthentication =
-        await googleSignInAccount.authentication;
-
-    final AuthCredential credential = GoogleAuthProvider.getCredential(
-        idToken: googleSignInAuthentication.idToken,
-        accessToken: googleSignInAuthentication.accessToken);
-
-    AuthResult result = await _auth.signInWithCredential(credential);
-    FirebaseUser userDetails = result.user;
-
-    if (result == null) {
-    } else {
-      Navigator.push(context, MaterialPageRoute(builder: (context) => Chat()));
     }
   }
