+++ FlutterChatAppTutorial/lib/feature/presentation/ui/pages/signin.dart
@@ -1,0 +1,183 @@
+import 'package:chatapp/feature/data/services/auth.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:flutter/material.dart';
+
+import '../../../../core/shared_preferences/helperfunctions.dart';
+
+import '../../../data/database/database.dart';
+import '../widget/widget.dart';
+import 'chatrooms.dart';
+
+class SignIn extends StatefulWidget {
+  final Function toggleView;
+
+  SignIn(this.toggleView);
+
+  @override
+  _SignInState createState() => _SignInState();
+}
+
+class _SignInState extends State<SignIn> {
+  TextEditingController emailEditingController = new TextEditingController();
+  TextEditingController passwordEditingController = new TextEditingController();
+
+  AuthService authService = new AuthService();
+
+  final formKey = GlobalKey<FormState>();
+
+  bool isLoading = false;
+
+  signIn() async {
+    if (formKey.currentState.validate()) {
+      setState(() {
+        isLoading = true;
+      });
+
+      await authService
+          .signInWithEmailAndPassword(
+              emailEditingController.text, passwordEditingController.text)
+          .then((result) async {
+        if (result != null) {
+          QuerySnapshot userInfoSnapshot =
+              await DatabaseMethods().getUserInfo(emailEditingController.text);
+
+          HelperFunctions.saveUserLoggedInSharedPreference(true);
+          HelperFunctions.saveUserNameSharedPreference(
+              userInfoSnapshot.documents[0].data["userName"]);
+          HelperFunctions.saveUserEmailSharedPreference(
+              userInfoSnapshot.documents[0].data["userEmail"]);
+
+          Navigator.pushReplacement(
+              context, MaterialPageRoute(builder: (context) => ChatRoom()));
+        } else {
+          setState(() {
+            isLoading = false;
+            //show snackbar
+          });
+        }
+      });
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: appBarMain(context),
+      body: isLoading
+          ? Container(
+              child: Center(child: CircularProgressIndicator()),
+            )
+          : Container(
+              padding: EdgeInsets.symmetric(horizontal: 24),
+              child: Column(
+                children: [
+                  Spacer(),
+                  Form(
+                    key: formKey,
+                    child: Column(
+                      children: [
+                        TextFormField(
+                          validator: (val) {
+                            return RegExp(
+                                        r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9]+\.[a-zA-Z]+")
+                                    .hasMatch(val)
+                                ? null
+                                : "Please Enter Correct Email";
+                          },
+                          controller: emailEditingController,
+                          style: simpleTextStyle(),
+                          decoration: textFieldInputDecoration("email"),
+                        ),
+                        TextFormField(
+                          obscureText: true,
+                          validator: (val) {
+                            return val.length >= 6
+                                ? null
+                                : "Enter Password 6+ characters";
+                          },
+                          style: simpleTextStyle(),
+                          controller: passwordEditingController,
+                          decoration: textFieldInputDecoration("password"),
+                        ),
+                      ],
+                    ),
+                  ),
+                  SizedBox(
+                    height: 16,
+                  ),
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.end,
+                    children: [
+                      GestureDetector(
+                        onTap: () {},
+                        child: Container(
+                            padding: EdgeInsets.symmetric(
+                                horizontal: 16, vertical: 8),
+                            child: Text(
+                              "Forgot Password?",
+                              style: simpleTextStyle(),
+                            )),
+                      )
+                    ],
+                  ),
+                  SizedBox(
+                    height: 16,
+                  ),
+                  GestureDetector(
+                    onTap: () {
+                      signIn();
+                    },
+                    child: Container(
+                      padding: EdgeInsets.symmetric(vertical: 16),
+                      decoration: BoxDecoration(
+                          borderRadius: BorderRadius.circular(30),
+                          gradient: LinearGradient(
+                            colors: [
+                              const Color(0xff007EF4),
+                              const Color(0xff2A75BC)
+                            ],
+                          )),
+                      width: MediaQuery.of(context).size.width,
+                      child: Text(
+                        "Sign In",
+                        style: biggerTextStyle(),
+                        textAlign: TextAlign.center,
+                      ),
+                    ),
+                  ),
+                  SizedBox(
+                    height: 16,
+                  ),
+                  SizedBox(
+                    height: 16,
+                  ),
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      Text(
+                        "Don't have account? ",
+                        style: simpleTextStyle(),
+                      ),
+                      GestureDetector(
+                        onTap: () {
+                          widget.toggleView();
+                        },
+                        child: Text(
+                          "Register now",
+                          style: TextStyle(
+                              color: Colors.white,
+                              fontSize: 16,
+                              decoration: TextDecoration.underline),
+                        ),
+                      ),
+                    ],
+                  ),
+                  SizedBox(
+                    height: 50,
+                  )
+                ],
+              ),
+            ),
+    );
+  }
+}
