+++ FlutterChatAppTutorial/lib/feature/presentation/ui/pages/signup.dart
@@ -1,59 +1,58 @@
-import 'package:chatapp/helper/helperfunctions.dart';
-import 'package:chatapp/helper/theme.dart';
-import 'package:chatapp/services/auth.dart';
-import 'package:chatapp/services/database.dart';
-import 'package:chatapp/views/chatrooms.dart';
-import 'package:chatapp/views/forgot_password.dart';
-import 'package:chatapp/widget/widget.dart';
-import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:chatapp/feature/data/services/auth.dart';
 import 'package:flutter/material.dart';
 
-class SignIn extends StatefulWidget {
+import '../../../../core/shared_preferences/helperfunctions.dart';
+
+import '../../../data/database/database.dart';
+import '../widget/widget.dart';
+import 'chatrooms.dart';
+
+class SignUp extends StatefulWidget {
   final Function toggleView;
 
-  SignIn(this.toggleView);
+  SignUp(this.toggleView);
 
   @override
-  _SignInState createState() => _SignInState();
+  _SignUpState createState() => _SignUpState();
 }
 
-class _SignInState extends State<SignIn> {
+class _SignUpState extends State<SignUp> {
   TextEditingController emailEditingController = new TextEditingController();
   TextEditingController passwordEditingController = new TextEditingController();
+  TextEditingController usernameEditingController = new TextEditingController();
 
   AuthService authService = new AuthService();
+  DatabaseMethods databaseMethods = new DatabaseMethods();
 
   final formKey = GlobalKey<FormState>();
-
   bool isLoading = false;
 
-  signIn() async {
+  singUp() async {
     if (formKey.currentState.validate()) {
       setState(() {
         isLoading = true;
       });
 
       await authService
-          .signInWithEmailAndPassword(
+          .signUpWithEmailAndPassword(
               emailEditingController.text, passwordEditingController.text)
-          .then((result) async {
-        if (result != null)  {
-          QuerySnapshot userInfoSnapshot =
-              await DatabaseMethods().getUserInfo(emailEditingController.text);
+          .then((result) {
+        if (result != null) {
+          Map<String, String> userDataMap = {
+            "userName": usernameEditingController.text,
+            "userEmail": emailEditingController.text
+          };
+
+          databaseMethods.addUserInfo(userDataMap);
 
           HelperFunctions.saveUserLoggedInSharedPreference(true);
           HelperFunctions.saveUserNameSharedPreference(
-              userInfoSnapshot.documents[0].data["userName"]);
+              usernameEditingController.text);
           HelperFunctions.saveUserEmailSharedPreference(
-              userInfoSnapshot.documents[0].data["userEmail"]);
+              emailEditingController.text);
 
           Navigator.pushReplacement(
               context, MaterialPageRoute(builder: (context) => ChatRoom()));
-        } else {
-          setState(() {
-            isLoading = false;
-            //show snackbar
-          });
         }
       });
     }
@@ -65,7 +64,9 @@ class _SignInState extends State<SignIn> {
       appBar: appBarMain(context),
       body: isLoading
           ? Container(
-              child: Center(child: CircularProgressIndicator()),
+              child: Center(
+                child: CircularProgressIndicator(),
+              ),
             )
           : Container(
               padding: EdgeInsets.symmetric(horizontal: 24),
@@ -77,27 +78,37 @@ class _SignInState extends State<SignIn> {
                     child: Column(
                       children: [
                         TextFormField(
+                          style: simpleTextStyle(),
+                          controller: usernameEditingController,
+                          validator: (val) {
+                            return val.isEmpty || val.length < 3
+                                ? "Enter Username 3+ characters"
+                                : null;
+                          },
+                          decoration: textFieldInputDecoration("username"),
+                        ),
+                        TextFormField(
+                          controller: emailEditingController,
+                          style: simpleTextStyle(),
                           validator: (val) {
                             return RegExp(
                                         r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9]+\.[a-zA-Z]+")
                                     .hasMatch(val)
                                 ? null
-                                : "Please Enter Correct Email";
+                                : "Enter correct email";
                           },
-                          controller: emailEditingController,
-                          style: simpleTextStyle(),
                           decoration: textFieldInputDecoration("email"),
                         ),
                         TextFormField(
                           obscureText: true,
-                          validator: (val) {
-                            return val.length >=6
-                                ? null
-                                : "Enter Password 6+ characters";
-                          },
                           style: simpleTextStyle(),
-                          controller: passwordEditingController,
                           decoration: textFieldInputDecoration("password"),
+                          controller: passwordEditingController,
+                          validator: (val) {
+                            return val.length < 6
+                                ? "Enter Password 6+ characters"
+                                : null;
+                          },
                         ),
                       ],
                     ),
@@ -105,32 +116,9 @@ class _SignInState extends State<SignIn> {
                   SizedBox(
                     height: 16,
                   ),
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.end,
-                    children: [
-                      GestureDetector(
-                        onTap: () {
-                          Navigator.push(
-                              context,
-                              MaterialPageRoute(
-                                  builder: (context) => ForgotPassword()));
-                        },
-                        child: Container(
-                            padding: EdgeInsets.symmetric(
-                                horizontal: 16, vertical: 8),
-                            child: Text(
-                              "Forgot Password?",
-                              style: simpleTextStyle(),
-                            )),
-                      )
-                    ],
-                  ),
-                  SizedBox(
-                    height: 16,
-                  ),
                   GestureDetector(
                     onTap: () {
-                      signIn();
+                      singUp();
                     },
                     child: Container(
                       padding: EdgeInsets.symmetric(vertical: 16),
@@ -144,7 +132,7 @@ class _SignInState extends State<SignIn> {
                           )),
                       width: MediaQuery.of(context).size.width,
                       child: Text(
-                        "Sign In",
+                        "Sign Up",
                         style: biggerTextStyle(),
                         textAlign: TextAlign.center,
                       ),
@@ -153,19 +141,6 @@ class _SignInState extends State<SignIn> {
                   SizedBox(
                     height: 16,
                   ),
-                  /*Container(
-                    padding: EdgeInsets.symmetric(vertical: 16),
-                    decoration: BoxDecoration(
-                        borderRadius: BorderRadius.circular(30),
-                        color: Colors.white),
-                    width: MediaQuery.of(context).size.width,
-                    child: Text(
-                      "Sign In with Google",
-                      style:
-                          TextStyle(fontSize: 17, color: CustomTheme.textColor),
-                      textAlign: TextAlign.center,
-                    ),
-                  ),*/
                   SizedBox(
                     height: 16,
                   ),
@@ -173,7 +148,7 @@ class _SignInState extends State<SignIn> {
                     mainAxisAlignment: MainAxisAlignment.center,
                     children: [
                       Text(
-                        "Don't have account? ",
+                        "Already have an account? ",
                         style: simpleTextStyle(),
                       ),
                       GestureDetector(
@@ -181,7 +156,7 @@ class _SignInState extends State<SignIn> {
                           widget.toggleView();
                         },
                         child: Text(
-                          "Register now",
+                          "SignIn now",
                           style: TextStyle(
                               color: Colors.white,
                               fontSize: 16,
@@ -197,5 +172,6 @@ class _SignInState extends State<SignIn> {
               ),
             ),
     );
+    ;
   }
 }
