+++ FlutterChatAppTutorial/lib/main.dart
@@ -1,7 +1,8 @@
-import 'package:chatapp/helper/authenticate.dart';
-import 'package:chatapp/helper/helperfunctions.dart';
-import 'package:chatapp/views/chatrooms.dart';
 import 'package:flutter/material.dart';
+
+import 'core/helper/authenticate.dart';
+import 'core/shared_preferences/helperfunctions.dart';
+import 'feature/presentation/ui/pages/chatrooms.dart';
 
 void main() {
   runApp(MyApp());
@@ -14,7 +15,6 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-
   bool userIsLoggedIn;
 
   @override
@@ -24,9 +24,9 @@ class _MyAppState extends State<MyApp> {
   }
 
   getLoggedInState() async {
-    await HelperFunctions.getUserLoggedInSharedPreference().then((value){
+    await HelperFunctions.getUserLoggedInSharedPreference().then((value) {
       setState(() {
-        userIsLoggedIn  = value;
+        userIsLoggedIn = value;
       });
     });
   }
@@ -43,12 +43,15 @@ class _MyAppState extends State<MyApp> {
         fontFamily: "OverpassRegular",
         visualDensity: VisualDensity.adaptivePlatformDensity,
       ),
-      home: userIsLoggedIn != null ?  userIsLoggedIn ? ChatRoom() : Authenticate()
+      home: userIsLoggedIn != null
+          ? userIsLoggedIn
+              ? ChatRoom()
+              : Authenticate()
           : Container(
-        child: Center(
-          child: Authenticate(),
-        ),
-      ),
+              child: Center(
+                child: Authenticate(),
+              ),
+            ),
     );
   }
 }
