--- FlutterChatAppTutorial/lib/views/forgot_password.dart
@@ -1,13 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ForgotPassword extends StatefulWidget {
-  @override
-  _ForgotPasswordState createState() => _ForgotPasswordState();
-}
-
-class _ForgotPasswordState extends State<ForgotPassword> {
-  @override
-  Widget build(BuildContext context) {
-    return Container();
-  }
-}
